refactor: improve scaffolding structure for common components in the EntityPropsPanel template
